canibalize BayesianEstimator by ModelBuilder
